refactor(rack): RackBuilderDialog-Modell nach rackBuilderModel.ts (Phase 5)
- Draft-Typen, 19"-Konstanten und reine Transform-Helfer (parseUnits,
  toPlacement, normalizeDraft, draftFromPreset, …) aus der 2.8k-Zeilen-
  Komponente ausgelagert (RackBuilderDialog 2864 → 2614 Zeilen)
- reines Verschieben: Markup + Verhalten unverändert, tsc verifiziert
- tiefere JSX/Hook-Zerlegung bewusst als laufzeit-verifizierter Folge-
  schritt dokumentiert (RISIKO ohne interaktiven Test)
- tsc + build grün, lint unverändert (124/18)
